+++ rs-admin/bootstrap-administration/src/test/java/fr/cnes/regards/microservice/modules/test/role/rest/RoleControllerIT.java
@@ -83,7 +83,7 @@ public class RoleControllerIT extends RegardsIntegrationTest {
 
     @Test
     public void createRole() {
-        Role newRole = new Role(15464, "new role", null, null, null);
+        Role newRole = new Role(15464L, "new role", null, null, null);
 
         ResponseEntity<Role> response = restTemplate.postForEntity(apiRoles, newRole, Role.class);
         assertEquals(HttpStatus.CREATED, response.getStatusCode());
@@ -94,7 +94,7 @@ public class RoleControllerIT extends RegardsIntegrationTest {
 
     @Test
     public void retrieveRole() {
-        Integer roleId = 0;
+        Long roleId = 0L;
         assertFalse(!this.roleService_.existRole(roleId));
 
         ParameterizedTypeReference<Role> typeRef = new ParameterizedTypeReference<Role>() {
@@ -102,7 +102,7 @@ public class RoleControllerIT extends RegardsIntegrationTest {
         ResponseEntity<Role> response = restTemplate.exchange(this.apiRolesId, HttpMethod.GET, null, typeRef, roleId);
         assertEquals(HttpStatus.OK, response.getStatusCode());
 
-        Integer wrongRoleId = 46453;
+        Long wrongRoleId = 46453L;
         assertFalse(this.roleService_.existRole(wrongRoleId));
         ResponseEntity<Object> responseNotFound = restTemplate.exchange(this.apiRolesId, HttpMethod.GET, null,
                                                                         Object.class, wrongRoleId);
@@ -111,7 +111,7 @@ public class RoleControllerIT extends RegardsIntegrationTest {
 
     @Test
     public void updateRole() {
-        Integer roleId = 0;
+        Long roleId = 0L;
         assertThat(roleService_.existRole(roleId));
         Role updated = roleService_.retrieveRole(roleId);
         updated.setName("newName");
@@ -123,7 +123,7 @@ public class RoleControllerIT extends RegardsIntegrationTest {
         assertEquals(HttpStatus.OK, response.getStatusCode());
 
         // if that's not the same functional ID and the parameter is valid:
-        Integer notSameID = 41554;
+        Long notSameID = 41554L;
         Role notUpdated = new Role(notSameID, null, null, null, null);
         HttpEntity<Role> requestOperationNotAllowed = new HttpEntity<>(notUpdated);
         ResponseEntity<Void> responseOperationNotAllowed = restTemplate
@@ -133,7 +133,7 @@ public class RoleControllerIT extends RegardsIntegrationTest {
 
     @Test
     public void removeRole() {
-        Integer roleId = 0;
+        Long roleId = 0L;
         ParameterizedTypeReference<Void> typeRef = new ParameterizedTypeReference<Void>() {
         };
         ResponseEntity<Void> response = restTemplate.exchange(this.apiRolesId, HttpMethod.DELETE, null, typeRef,
@@ -143,7 +143,7 @@ public class RoleControllerIT extends RegardsIntegrationTest {
 
     @Test
     public void retrieveRoleResourcesAccessList() {
-        Integer roleId = 0;
+        Long roleId = 0L;
         ParameterizedTypeReference<List<ResourcesAccess>> typeRef = new ParameterizedTypeReference<List<ResourcesAccess>>() {
         };
         ResponseEntity<List<ResourcesAccess>> response = restTemplate.exchange(apiRolesPermissions, HttpMethod.GET,
@@ -153,11 +153,11 @@ public class RoleControllerIT extends RegardsIntegrationTest {
 
     @Test
     public void updateRoleResourcesAccess() {
-        Integer roleId = 0;
+        Long roleId = 0L;
 
         List<ResourcesAccess> newPermissionList = new ArrayList<>();
-        newPermissionList.add(new ResourcesAccess(463, "new", "new", "new", HttpVerb.PUT));
-        newPermissionList.add(new ResourcesAccess(350, "neww", "neww", "neww", HttpVerb.DELETE));
+        newPermissionList.add(new ResourcesAccess(463L, "new", "new", "new", HttpVerb.PUT));
+        newPermissionList.add(new ResourcesAccess(350L, "neww", "neww", "neww", HttpVerb.DELETE));
 
         ParameterizedTypeReference<Void> typeRef = new ParameterizedTypeReference<Void>() {
         };
@@ -169,7 +169,7 @@ public class RoleControllerIT extends RegardsIntegrationTest {
 
     @Test
     public void clearRoleResourcesAccess() {
-        Integer roleId = 0;
+        Long roleId = 0L;
 
         ParameterizedTypeReference<Void> typeRef = new ParameterizedTypeReference<Void>() {
         };
@@ -180,7 +180,7 @@ public class RoleControllerIT extends RegardsIntegrationTest {
 
     @Test
     public void retrieveRoleProjectUserList() {
-        Integer roleId = 0;
+        Long roleId = 0L;
 
         ParameterizedTypeReference<List<ProjectUser>> typeRef = new ParameterizedTypeReference<List<ProjectUser>>() {
         };
